refactor(exception): exception의 양이 많아짐에 따라  도메인별 분류
