+++ sgh/src/main/java/com/growing/sgh/advice/ExceptionAdvice.java
@@ -1,7 +1,15 @@
 package com.growing.sgh.advice;
 
-import com.growing.sgh.exception.*;
 import com.growing.sgh.common.response.Response;
+import com.growing.sgh.exception.item.ItemImgFileNotFoundException;
+import com.growing.sgh.exception.item.ItemNotFoundException;
+import com.growing.sgh.exception.item.RegisterImgNotExistsException;
+import com.growing.sgh.exception.member.EmailAlreadyExistsException;
+import com.growing.sgh.exception.member.MemberNotFoundException;
+import com.growing.sgh.exception.member.NicknameAlreadyExistsException;
+import com.growing.sgh.exception.member.UsernameAlreadyExistsException;
+import com.growing.sgh.exception.order.OutOfStockException;
+import com.growing.sgh.exception.order.SoldOutItemException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
