+++ sgh/src/main/java/com/growing/sgh/domain/item/controller/ItemController.java
@@ -7,9 +7,8 @@ import com.growing.sgh.domain.item.dto.ItemSearchDto;
 import com.growing.sgh.domain.item.entity.Item;
 import com.growing.sgh.domain.item.service.ItemImgService;
 import com.growing.sgh.domain.item.service.ItemService;
-import com.growing.sgh.exception.RegisterImgNotExistsException;
+import com.growing.sgh.exception.item.RegisterImgNotExistsException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
