+++ sgh/src/main/java/com/growing/sgh/domain/order/service/OrderService.java
@@ -13,7 +13,11 @@ import com.growing.sgh.domain.order.dto.OrderItemDto;
 import com.growing.sgh.domain.order.entity.Order;
 import com.growing.sgh.domain.order.entity.OrderItem;
 import com.growing.sgh.domain.order.repository.OrderRepository;
-import com.growing.sgh.exception.*;
+import com.growing.sgh.exception.item.ItemNotFoundException;
+import com.growing.sgh.exception.member.MemberDoesNotMatchException;
+import com.growing.sgh.exception.member.MemberNotFoundException;
+import com.growing.sgh.exception.order.OrderNotFoundException;
+import com.growing.sgh.exception.order.SoldOutItemException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
