+++ sgh/src/main/java/com/growing/sgh/exception/item/ItemImgFileNotFoundException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.item;
 
 public class ItemImgFileNotFoundException extends RuntimeException{
 }
