+++ sgh/src/main/java/com/growing/sgh/exception/member/EmailAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.member;
 
 public class EmailAlreadyExistsException extends RuntimeException{
     public EmailAlreadyExistsException(String message) {
