+++ sgh/src/main/java/com/growing/sgh/exception/member/NicknameAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.member;
 
 public class NicknameAlreadyExistsException extends RuntimeException{
     public NicknameAlreadyExistsException(String message) {
