+++ sgh/src/main/java/com/growing/sgh/exception/member/UsernameAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.member;
 
 public class UsernameAlreadyExistsException extends RuntimeException {
     public UsernameAlreadyExistsException(String message){
