+++ sgh/src/main/java/com/growing/sgh/exception/order/OrderNotFoundException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.order;
 
 public class OrderNotFoundException extends RuntimeException{
     public OrderNotFoundException() {
