+++ sgh/src/main/java/com/growing/sgh/exception/order/OutOfStockException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.order;
 
 public class OutOfStockException extends RuntimeException{
     public OutOfStockException() {
