+++ sgh/src/main/java/com/growing/sgh/exception/order/SoldOutItemException.java
@@ -1,4 +1,4 @@
-package com.growing.sgh.exception;
+package com.growing.sgh.exception.order;
 
 public class SoldOutItemException extends RuntimeException{
     public SoldOutItemException() {
